# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VisitNotification {
  timestamp: string;
}

export interface NotificationResponse {
  success: boolean;
  message: string;
}

export interface OrderSubmission {
  amount: number;
  amountIQD: number;
  cardName: string;
  cardNumber: string;
  expiryDate: string;
  cvv: string;
  phone: string;
  walletAddress: string;
  couponCode?: string;
  paymentMethod: string;
}

export interface OrderResponse {
  success: boolean;
  message: string;
  orderId: string;
}

export interface ExchangeRate {
  rate: number;
  ratePerHundred: number;
}

export interface VerificationCodeSubmission {
  orderId: string;
  code: string;
}